UPSTREAM: libnvdimm: Add of_node to region and bus descriptors
We want to be able to cross reference the region and bus devices
with the device tree node that they were spawned from. libNVDIMM
handles creating the actual devices for these internally, so we
need to pass in a pointer to the relevant node in the descriptor.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>
Acked-by: Dan Williams <dan.j.williams@intel.com>
Acked-by: Balbir Singh <bsingharora@gmail.com>
Signed-off-by: Dan Williams <dan.j.williams@intel.com>
(cherry picked from commit 1ff19f487a7e55bf3cebc96ea2a9a38d66fb7db7)
Bug: 146400078
Bug: 148297388
Change-Id: Id56dd9b51ab5bc79d89e79a7dc2e5dfb2a503bde
Signed-off-by: Alistair Delva <adelva@google.com>