mm: Drop Google's picks and put do_wp_page back in line with stable
This causes TEEGRIS to crash out and die. The user-visible result is
that the lockscreen pin is claimed to be incorrect, making the device
unusable. The device may be unlocked once upon boot, if done quickly
enough, but not after relocking or sitting long enough after boot.

This is a squash of:
* Revert "UPSTREAM: mm: reuse only-pte-mapped KSM page in do_wp_page()"
  This reverts commit 8948d08467565826915b47996586a5b44c2038f0.
* Revert "UPSTREAM: mm/ksm: Remove reuse_ksm_page()"
  This reverts commit d094baca465dae16f723e816517061956cb6d203.
* Revert "BACKPORT: mm: do_wp_page() simplification"
  This reverts commit b6b3d960cbc4e1414114e8f50e16a2f408e15320.
* Revert "UPSTREAM: mm: fix misplaced unlock_page in do_wp_page()"
  This reverts commit 249de5f1ab785003a26a13f1cc2818d417322cf5.

Signed-off-by: Kevin F. Haggerty <haggertk@lineageos.org>
Change-Id: I83980aea81a0b667068acdf6ffc7d698ff1d51bf